Use expression and virtual columns to express projects and aggregates. Refactor the checks of aggs and projects, now checkis done via translation method. Adding more tests for new capabilities.

+++ druid/src/main/java/org/apache/calcite/adapter/druid/DruidConnectionImpl.java
@@ -347,7 +347,7 @@ class DruidConnectionImpl implements DruidConnection {
     switch (token) {
     case VALUE_NUMBER_INT:
       if (type == null) {
-        type = ColumnMetaData.Rep.INTEGER;
+        type = ColumnMetaData.Rep.LONG;
       }
       // fall through
     case VALUE_NUMBER_FLOAT:
